Add -ignore-merge command flag
This flag can be used to ignore merge commits as they are not signed
by gitlab.

Even if we could do automatic signing it would defeat the proposal of
pistis, as we want to use signatures as a notarisation device.

For this to be effective we need to change the order in which we parse
the tree to BFS with git.LogOrderBSF.

That way we ensure that we validate the content of the git merge
before proceeding further down history. And it is this content that
will be signed.

https://en.wikipedia.org/wiki/Breadth-first_search