Read usernames from GitLab
To reduce the need for redundantly stored information in order to use
Pistis, optionally read the usernames, which are used to fetch public
PGP keys from GitLab, from GitLab as well.
As GitLab allows querying of user data based on email addresses only
for authorized users, the new functionality is only enabled if a
respective API token is found under "GITLAB_TOKEN" in the environment.
Users in the existing CODEOWNERS_USERNAME file will still be
considered.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>