Commit exclusions in username iteration
The commit exclusions using the TRUSTED_COMMITS file were only considered
for validation of commits (the second history iteration), whilst the
scan for GitLab usernames always parsed the whole history. Scanning for
usernames of commits which are not going to be checked is useless - make
the program execution more efficient by only scanning for GitLab
usernames of commits which are also going to be checked for authenticity.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>